Add Event_Get_T5R2
The T5R2 property is set for synaptics image sensor touchpads.
These touchpads support a kind of multitouch where up to 5 fingers are
tracked, but only two of these positions are reported, the first used
slot, and the last used slot.

The other slots (between first and last) are considered "hidden".
They are reported with a tracking_id != -1, however, the position data
within the slot must not be considered valid.

In summary:
 * The number of fingers on the pad is the number of slots with
   tracking_id != -1.
 * The first slot with tracking_id != -1 contains position data for one
   finger.
 * The last slot with tracking_id != -1 contains position data for the
   other finger.
 * All other slots with tracking_id != -1 do not contain valid position
   data.

Note: T5R2 is not yet included in the upstream kernel.

BUG=chromium-os:17288
TEST=none

Change-Id: I7bf690520e8a8d42ba46ca98d04197e524bc7c5c