Implement gestures lib GesturesProp interface
Gestures lib can create XI properties of type Int, Short, Bool, Real or
String.  Each property must have a unique name, which is converted to an
Atom which is managed by the server.

When creating a property, a pointer to a variable of the appropriate type
should be provided.  If NULL is provided, the property is ReadOnly.
Otherwise, when the server updates a property, the passed-in variable
is automatically updated as well.

Unlike traditional XI properties, each Gesture property can contain
exactly one value.  Its value may be initialized by an identically named
(and typed) Option value in an xorg.conf snippet file.  If no such entry
is found, then the default value is as passed in to the property creation
function.

BUG=chromium-os:17991
TEST=compile w/ no errors;
     props with correct names accessible w/ xinput
     props initialized with .conf snippet values if present

Change-Id: I7389fd671529ff23bcc3185a91d2fe804a3e104d