Include xorg-server.h before any other X header files
xorg-server.h contains various defines that are used by other X header
files.

In particular, on 64-bit machines it defines _XSERVER64, which is used
to determine whether the "Atom" type is "unsigned long" or "CARD32".
This has a drastic affect on the layout of various internal structs shared
between the driver and the x server.

BUG=chromium-os:22902
TEST=build with:
 USE="cmt" INPUT_DEVICES="cmt evdev" build_packages --board=amd64-generic
 * cmt loads properly, detects trackpad, correctly initializes based on
   kernel driver properties, as reported in /var/log/Xorg.0.log
 * Verify pointer motion, click, tap, scroll, etc. all perform as expected

Change-Id: I7c1df73c09f7fc33b849a7988756019817519cf6