Fix the pressure readings of semi_mt devices for gesture recognition.
The current cmt driver did not support semi-mt device well such as
Synaptics touchpad on Cr-48. The major issues are from the pressure
readings are not passed into the gesture library and its sensitivity.
For semi-MT, drivers do not report a 'per-contact' pressure
(ABS_MT_PRESSURE), but instead report a single pressure that is assumed
to be constant for all fingers (ABS_PRESSURE). This is unfortunate, but
the best that can be done given the limitations of semi-mt. We
need to retrieve the readings from ABS_PRESSURE events instead of
ABS_MT_PRESSURE for SEMI_MT devices.

BUG=chromium-os:24951
TEST=test the cmt on Cr48, and this change should not see anything
different from users.

Change-Id: I62578553ac4d1a735fef5ca9947c84b3eb5aebb0