Use monotonic timestamps for input events if available in kernel
This also requires that the timer provider use the same monotonic clock
for scheduling timer callbacks.  This is done using
clock_gettime(CLOCK_MONOTONIC, x).

Most of the complication of this patch involves routing the 'is_monontic'
indicator from the kernel to the timer callback.  To do this, the
previously unused GestureInterpreterSetTimerProvider(,, data) data field
is used to store a reference to the cmt device.

BUG=chromium-os:26493
TEST=on kernel with monotonic timestamp patch applied;
   add "-logverbose 10" to X invocation in /sbin/startx.sh
   from shell, tail -F /var/log/Xorg.0.log &
   move fingers on touchpad.
   Note that device events are timestamped with a monotonic clock
   (Timestamps will be close to, but not exactly equal to, X log times)


Change-Id: I23311b80293f84c0754c96b9c774587e60d3f38c