Store last timestamp as cmt data, not static
This patch cleans up the timestamp valuator implementation.
In particular, it stores the 'last timestamp', used for relative to
absolute timestamp conversion, as member variables of the cmt struct
instead of static variables.  In this way, a single driver can support
multiple instances.

Also, the Button valuators are explicitly posted as absolute events,
whereas the motion and scroll motion events are posted as relative
events.  It is unclear whether this has any observable affect or not.

Hopefully, the treatment of these valuators will become more rational
once ABI 12 is introduced with the forthcoming upgrade to xorg-server
1.12.

BUG=chromium-os:25181
TEST=builds clean; xinput test-xi2 shows timestamps for motion and
     button events with CMT.

Change-Id: I2e2a42a6f89c00b0ee5d5d414846956fdc44fe88