Fill pressure information for both slots on Cr48
On Cr48, the pressure event may not be informed right after a new slot
is reported, instead, the previous value of the new slot will be
reported. However, we have to use the same pressure value as the other slot
(i.e., the current pressure value) for the semi_mt device.

Therefore, the pressure value will be filled into both slots for every
ABS_PRESSURE event even if the slot is unused.

BUG=chromium-os:27520
TEST=test it on Cr48 with CMT.

Change-Id: I2d8a2da20b15c7e6e33bf796c34573fcc377b132