Synchronize cmt state with kernel evdev driver
A SYN_DROPPED event indicates the the kernel's internal event buffer
overflowed, and some events will not be delivered to userspace.
This can directly lead to 'ghost finger' (userspace thinking a finger
exists after it is no longer touching a pad), for example, if a
TRACKING_ID=-1 event is dropped for a given mt slot.

BUG=chromium-os:28997
TEST=run the following test on device and see if it is completed successfully.
  autotest platform_TrackpadSynDrop/control

Change-Id: Ice0ea3574b88635a86134e274e921978cd5d5142