refactoring: moved last ioctl calls from libevdev_event to libevdev.
Now all device(fd) accessing methods are gathered in libevdev.c and can
be easily mocked for regression testing.
Note: CMT was using the X.org input driver guidelines. New changes in the code
are written with the google style guidelines, so there is a style mismatch
in some cases.
This code will be moved to a separate library in upcoming CLs that will
completely follow the google guidelines.

BUG=chromium-os:31022
TEST=touch regression suite
Change-Id: I529cc3140d028be8e230b33e94eb7d61279fcc20