refactoring: moved device info ioctl's to libevdev
During initialization in libevdev_event.c the kernel is queried
for device information using ioctl's. This code has been moved to
libevdev. This allows nitrous to reuse the ioctl code and the
regression test suite to easily mock the ioctl code with fake data.

The remaining ioctl's will be moved to libevdev in the following
CLs.

BUG=chromium-os:31022
TEST=touch regression suite

Change-Id: I3cb2adf81e8b3888822df44112d5cd3e1220ff40