Pass device class to gesture interpreter
The gesture interpreter needs device class to initialize itself.  While
we had once intended to set this information in xorg.conf and so gesture
interpreter may acquire it through properties, we decided that we would
let libevdev probe this and pass it on.  Thus, the gesture interpreter
initializer was added a new parameter for device class.

BUG=chromium-os:36634
TEST=On Snow, touchpad, Magic Trackpad, and generic mouse still work

Change-Id: I72d563e83a68264b02569edd2ca14d2ca6e9bc29
Reviewed-on: https://gerrit.chromium.org/gerrit/38696
Commit-Ready: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Tested-by: Che-Liang Chiou <clchiou@chromium.org>