Feature: Raw Touch Passthrough Mode
Adds the "Raw Touch Passthrough" property that activates a mode
which suppresses all mouse emulation events and passes through the
raw touch events from the evdev device.

When fingers are present when activating the raw mode, those fingers
are removed from the gesture interpretation and instead passed through
as new fingers.
When fingers are present when deactivating the raw mode, those fingers
are ignored by the gesture interpretation. That means a finger that
is present during the deactivation cannot be used for gesturing.

BUG=chromium:325285
TEST=manual testing on link

Change-Id: I4b67ab657dc4ad9705b28b9b009d6362fb14d5c7
Reviewed-on: https://chromium-review.googlesource.com/178600
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Commit-Queue: Dennis Kempin <denniskempin@google.com>
Tested-by: Dennis Kempin <denniskempin@google.com>